Update README.md
had to add a 2nd startup script for VNC to strip the INTERNAL IP properly - this may have come from using a larger machine type (8 CPU, 30 GB RAM) or a new line format from ifconfig that now lists the IP address as 'addr:x.x.x.x' instead of just 'x.x.x.x'

I also removed the ';' at the end of the startup script. Not sure if that got added somehow or if it prevented command from completing and starting the container. But in the prior run, the container didn't start running (which may have also been due to an improper add of 'COPY resolve/resolv.conf:/etc/resolv.conf' in the Docker image build file that I made on a prior run.